[CLOUD-3119] txrecovery module to run with python3 (#64)
