+++ os-eap-txnrecovery/bash/added/logging.properties
