+++ os-eap-txnrecovery/bash/module.yaml
@@ -1,10 +1,7 @@
 schema_version: 1
-name: os-eap-txnrecovery
+name: os-eap-txnrecovery.bash
 version: '1.0'
 description: Transaction recovery migration pod script package.
 execute:
 - script: install_as_root.sh
-packages:
-    install:
-        - python-enum34
-        - python-requests
+
