+++ os-eap-txnrecovery/run/python2/install_as_root.sh
@@ -1,0 +1,8 @@
+set -u
+set -e
+
+SCRIPT_DIR=$(dirname $0)
+
+cp "$SCRIPT_DIR"/*.py /opt/partition/
+
+chmod 755 /opt/partition/*
