+++ os-eap64-launch/module.yaml
@@ -8,7 +8,6 @@ modules:
   - name: jboss.container.maven.default.bash
   - name: os-eap-launch
   - name: os-eap-node-name
-  - name: os-eap-txnrecovery
   - name: os-eap-migration
 
 execute:
