+++ os-eap7-launch/module.yaml
@@ -7,7 +7,6 @@ modules:
   install:
   - name: os-eap-launch
   - name: os-eap-node-name
-  - name: os-eap-txnrecovery
   - name: os-eap-migration
 
 execute:
